requirements.txt
chromadb>=1.5.0
requests>=2.32.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0